# openapi.yml
openapi: "3.0.3"
info:
  title: "MyLittleShopping API"
  version: 0.0.1
paths:
...
  /products:
    get:
      description: "Get all products data according to your role as a user"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/product"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
    parameters:
      - in: query
        name: condition
        schema:
          type: string
      - in: query
        name: category
        schema:
          type: string
      - in: query
        name: tag
        schema:
          type: string
      - in: query
        name: filter
        schema:
          type: string
      - in: query
        name: sellerUsername
        schema:
          type: string
      - in: query
        name: lowerPrice
        schema:
          type: number
          multipleOf: 0.01
      - in: query
        name: higherPrice
        schema:
          type: number
          multipleOf: 0.01

    post:
      description: "Allow a user to create a product for sell"
      responses:
        201:
          description: "Product created"
          content:
            text/plain:
              schema:
                type: string
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                label:
                  type: string
                condition:
                  type: string
                  enum: [new, renovated, occasion]
                description:
                  type: string
                unitPrice:
                  type: number
                  multipleOf: 0.01
                availableQuantity:
                  type: integer
              required:
                - name
                - label
                - condition
                - description
                - unitPrice
                - availableQuantity

    put:
      description: "Allow a seller to modify one of his products"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                items:
                  allOf:
                    - $ref: "#/components/schemas/product"
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                categories:
                  type: string
                tags:
                  type: string
    delete:
      description: "Allow a seller to delete one of his products"
      responses:
        200:
          description: "Product deleted"
          content:
            text/plain:
              schema:
                type: string
        400:
          description: "Bad Json Format"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
              required:
                - name
...
components:
  schemas:
...
    product:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        label:
          type: "string"
        condition:
          type: "string"
          description: "Condition of a product"
          enum: ["new", "occasion", "renovated"]
        description:
          type: "string"
        unitPrice:
          type: "number"
          multipleOf: 0.01
        availableQuantity:
          type: "integer"
          format: "int64"
        sellerId:
          type: "integer"
          format: "int64"
        onSale:
          type: "boolean"
        averageRating:
          type: "integer"
          format: "int32"
          minimum: 0
          maximum: 5
        created_at:
          type: string
          format: date
        updated_at:
          type: string
          format: date
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT